requirements.txt
# Care Sync - Backend Dependencies
# Install with: pip3 install -r requirements.txt
# Or from backend directory: pip3 install -r backend/requirements.txt

# FastAPI web framework
fastapi==0.104.1

# ASGI server for FastAPI
uvicorn[standard]==0.24.0

# Data validation using Python type annotations
pydantic>=2.8.0

# OpenAI API client for GPT fact extraction
openai==1.3.5

# Environment variable management
python-dotenv==1.0.0

# Multipart form data support
python-multipart==0.0.6

# HTTP client for async requests
httpx==0.25.2
